book_store/book_service/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
pydantic
python-dotenv
fakeredis
redis
httpx